+++ kernel/plugin/rpc.go
@@ -24,7 +24,6 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/fastschema/qjs"
 	"github.com/gin-gonic/gin"
 	"github.com/gorilla/websocket"
 	"github.com/siyuan-note/logging"
@@ -318,118 +317,4 @@ func parseRpcRequests(body []byte) (requests []*JsonRpcInboundRequest, isBatch b
 	}
 	requests = append(requests, &request)
 	return requests, false, nil
-}
-
-// rpcParamsToJsValue converts Go RPC params to a *qjs.Value or []*qjs.Value for JS invocation.
-// It accepts various Go types (string, []byte, json.RawMessage, or any JSON-serializable struct) and returns an error if conversion fails.
-func rpcParamsToJsValue(ctx *qjs.Context, params any) (isArray bool, value *qjs.Value, array []*qjs.Value, err error) {
-	if params == nil {
-		return
-	}
-
-	var jsonStr string
-	switch v := params.(type) {
-	case string:
-		jsonStr = v
-	case *string:
-		if v == nil {
-			return
-		}
-		jsonStr = *v
-	case []byte:
-		jsonStr = string(v)
-	case json.RawMessage:
-		jsonStr = string(v)
-	case *[]byte:
-		if v == nil {
-			return
-		}
-		jsonStr = string(*v)
-	case *json.RawMessage:
-		if v == nil {
-			return
-		}
-		jsonStr = string(*v)
-	default:
-		b, marshalErr := json.Marshal(v)
-		if marshalErr != nil {
-			err = marshalErr
-			return
-		}
-		jsonStr = string(b)
-	}
-
-	isArray = isJsonArray(jsonStr) // quick check to determine if we should parse as array or single value
-	if isArray {
-		array, err = parseJsonArrayStringToJsValueArray(ctx, jsonStr)
-	} else {
-		value, err = parseJsonStringToJsValue(ctx, jsonStr)
-	}
-	return
-}
-
-// invokeRpcMethod calls the given JS function with the provided params and returns the result or error.
-func invokeRpcMethod(ctx *qjs.Context, rpcMethodName string, rpcMethod *qjs.Value, rpcParams any) (rpcResult any, rpcError *JsonRpcError) {
-	// Convert params to JS value
-	isArray, paramValue, paramArray, convertErr := rpcParamsToJsValue(ctx, rpcParams)
-	if convertErr != nil {
-		return nil, &JsonRpcError{
-			Code:    JsonRpcErrorCodeInternalError,
-			Message: fmt.Sprintf("convert params: %v", convertErr),
-		}
-	}
-
-	// Call the JS function using ctx.Invoke(fn, thisVal, args...)
-	var result *qjs.Value
-	var err error
-	if isArray {
-		result, err = ctx.Invoke(rpcMethod, ctx.Global(), paramArray...)
-	} else {
-		if paramValue == nil {
-			result, err = ctx.Invoke(rpcMethod, ctx.Global())
-		} else {
-			result, err = ctx.Invoke(rpcMethod, ctx.Global(), paramValue)
-		}
-	}
-	if err != nil {
-		return nil, &JsonRpcError{
-			Code:    JsonRpcErrorCodeInternalError,
-			Message: fmt.Sprintf("call %q: %v", rpcMethodName, err),
-		}
-	}
-
-	// Await if Promise
-	if result != nil && result.IsPromise() {
-		result, err = result.Await()
-		if err != nil {
-			return nil, &JsonRpcError{
-				Code:    JsonRpcErrorCodeInternalError,
-				Message: fmt.Sprintf("await %q: %v", rpcMethodName, err),
-			}
-		}
-	}
-
-	// If result is null or undefined, return nil (JSON-RPC allows result to be null, but we treat undefined as null for convenience)
-	if result == nil || result.IsNull() || result.IsUndefined() {
-		return nil, nil
-	}
-
-	// Convert JS result to Go via JSON round-trip using the built-in JSONStringify method.
-	jsonStr, stringifyErr := result.JSONStringify()
-	if stringifyErr != nil {
-		return nil, &JsonRpcError{
-			Code:    JsonRpcErrorCodeInternalError,
-			Message: fmt.Sprintf("stringify result: %v", stringifyErr),
-		}
-	}
-
-	// Unmarshal JSON string back to Go type (could be any JSON-serializable type: object, array, string, number, bool, or null)
-	if err = json.Unmarshal([]byte(jsonStr), &rpcResult); err != nil {
-		return nil, &JsonRpcError{
-			Code:    JsonRpcErrorCodeInternalError,
-			Message: fmt.Sprintf("unmarshal result: %v", err),
-		}
-	}
-
-	return
 }
